+++ client/src/App.js
@@ -28,7 +28,6 @@ function App() {
 
   return (
     <div className="App">
-      <ChakraProvider />
         {showNav && <NavBar />}
         <Routes>
           <Route path="/" element={<HomePage />} />
@@ -44,7 +43,6 @@ function App() {
           <Route path="/ResetPass" element={<Resetpass />} />
           <Route path="/About" element={<About />} />
         </Routes>
-      <ChakraProvider />
     </div>
   );
 }
