+++ client/src/Components/Products/arrayProducts.js
@@ -1,111 +1,163 @@
-export const array = 
-[
+export const array = [
 	{
-		nombre: "Hamburguesa de lentejas",
-		tipo: "Comida",
-		descripcion:
-			"Hamburguesa de lentejas, acompañada de lechuga, tomate, cebolla y mayonesa vegana.",
-		precio: 150,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+		id:1,
+		estado: "Pendiente",
+		importe: 54893632,
+		fecha: "12/12/2021",
+		productos: [
+			{
+				nombre: "Hamburguesa de lentejas",
+				tipo: "Comida",
+				descripcion:
+					"Hamburguesa de lentejas, acompañada de lechuga, tomate, cebolla y mayonesa vegana.",
+				precio: 150,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Pizza de verduras",
+				tipo: "Comida",
+				descripcion:
+					"Pizza vegana con una base de tomate y especias, cubierta de verduras frescas y queso vegano.",
+				precio: 200,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Sándwich de aguacate y tofu",
+				tipo: "Comida",
+				descripcion:
+					"Sándwich vegano con aguacate, tofu, lechuga y tomate, en pan integral.",
+				precio: 120,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Ensalada de quinoa",
+				tipo: "Comida",
+				descripcion:
+					"Ensalada vegana de quinoa con tomate, pepino, cebolla, aceitunas y aliño de limón.",
+				precio: 90,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+		],
 	},
+	
 	{
-		nombre: "Pizza de verduras",
-		tipo: "Comida",
-		descripcion:
-			"Pizza vegana con una base de tomate y especias, cubierta de verduras frescas y queso vegano.",
-		precio: 200,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+		id:2,
+		estado: "En proceso",
+		importe: 24558,
+		fecha: "15/1/2022",
+		productos: [
+			{
+				nombre: "Wrap de falafel",
+				tipo: "Comida",
+				descripcion:
+					"Wrap vegano con falafel, hummus, tomate, cebolla y lechuga, envuelto en tortilla integral.",
+				precio: 130,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Curry de garbanzos",
+				tipo: "Comida",
+				descripcion:
+					"Curry vegano de garbanzos con arroz integral, tomate, cebolla y especias.",
+				precio: 180,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Tacos de coliflor",
+				tipo: "Comida",
+				descripcion:
+					"Tacos veganos de coliflor, con guacamole, tomate, cebolla y cilantro.",
+				precio: 160,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Ensalada de tofu y frutos secos",
+				tipo: "Comida",
+				descripcion:
+					"Ensalada fresca de tofu, lechuga, zanahoria y frutos secos con aderezo de mostaza y miel.",
+				precio: 120,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Curry de calabaza y garbanzos",
+				tipo: "Comida",
+				descripcion:
+					"Un delicioso curry vegano de calabaza y garbanzos con arroz integral y un toque de cilantro fresco.",
+				precio: 180,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Hamburguesa de tofu y remolacha",
+				tipo: "Comida",
+				descripcion:
+					"Acompañada de lechuga, tomate y mayonesa vegana en pan integral.",
+				precio: 150,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+			{
+				nombre: "Tarta de chocolate y almendras",
+				tipo: "Comida",
+				descripcion:
+					"Un postre vegano irresistible, con un toque de vainilla y servida con una bola de helado vegano.",
+				precio: 150,
+				stock: 88,
+				cantidad: 5,
+				fecha: "12/12/2022",
+				importe: 454545,
+				estado: "Pendiente",
+				imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
+			},
+		],
 	},
-	{
-		nombre: "Sándwich de aguacate y tofu",
-		tipo: "Comida",
-		descripcion:
-			"Sándwich vegano con aguacate, tofu, lechuga y tomate, en pan integral.",
-		precio: 120,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Ensalada de quinoa",
-		tipo: "Comida",
-		descripcion:
-			"Ensalada vegana de quinoa con tomate, pepino, cebolla, aceitunas y aliño de limón.",
-		precio: 90,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Wrap de falafel",
-		tipo: "Comida",
-		descripcion:
-			"Wrap vegano con falafel, hummus, tomate, cebolla y lechuga, envuelto en tortilla integral.",
-		precio: 130,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Curry de garbanzos",
-		tipo: "Comida",
-		descripcion:
-			"Curry vegano de garbanzos con arroz integral, tomate, cebolla y especias.",
-		precio: 180,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Tacos de coliflor",
-		tipo: "Comida",
-		descripcion:
-			"Tacos veganos de coliflor, con guacamole, tomate, cebolla y cilantro.",
-		precio: 160,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Sopa de lentejas y espinacas",
-		tipo: "Comida",
-		descripcion:
-			"Una sopa cremosa de lentejas y espinacas, con un toque de jengibre y comino. Servida con pan integral.",
-		precio: 100,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Ensalada de tofu y frutos secos",
-		tipo: "Comida",
-		descripcion:
-			"Ensalada fresca de tofu, lechuga, zanahoria y frutos secos con aderezo de mostaza y miel.",
-		precio: 120,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Curry de calabaza y garbanzos",
-		tipo: "Comida",
-		descripcion:
-			"Un delicioso curry vegano de calabaza y garbanzos con arroz integral y un toque de cilantro fresco.",
-		precio: 180,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Hamburguesa de tofu y remolacha",
-		tipo: "Comida",
-		descripcion:
-			"Acompañada de lechuga, tomate y mayonesa vegana en pan integral.",
-		precio: 150,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
-	{
-		nombre: "Tarta de chocolate y almendras",
-		tipo: "Comida",
-		descripcion:
-			"Un postre vegano irresistible, con un toque de vainilla y servida con una bola de helado vegano.",
-		precio: 150,
-		stock: 88,
-		imagen: "https://cdn.britannica.com/24/174524-050-A851D3F2/Oranges.jpg",
-	},
+
 ];
