+++ client/src/redux/actions/actions.js
@@ -143,24 +143,6 @@ export const changePage = (number) => {
   return { type: SET_PAGE, payload: number };
 };
 
-// ACTIONS NECESARIAS:
-
-//* getAllProducts
-//* getProductById
-//* createOrder
-//* getOrders
-//* getOrderDetail
-//! deleteOrder
-//! updateOrder
-//! getClientData
-//! updateClientData
-//! deleteClient
-
-// order = {
-//   precioTotal: 1000,
-//   cliente_id: 1,
-//   productos: [1,3,2],
-// };
 
 /* CREAR ORDEN */
 export const createOrder = (order) => {
